@@ -8,6 +8,7 @@
     "clean": "rm -rf ./lib",
     "build-modules": "babel ./src --out-dir ./lib",
     "build": "npm run clean && npm run build-modules",
+    "prepublish": "npm run build",
     "test": "jest --config ./.jestrc",
     "cover": "jest --config ./.jestrc --coverage",
     "ci": "npm run lint && npm run cover"
@@ -22,6 +23,9 @@
   ],
   "author": "Teddy Ortega <tortega128@gmail.com>",
   "license": "MIT",
+  "dependencies": {
+    "prop-types": "^15.6.0"
+  },
   "devDependencies": {
     "babel-cli": "^6.23.0",
     "babel-core": "^6.23.1",
@@ -29,7 +33,8 @@
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.23.0",
     "babel-preset-stage-0": "^6.22.0",
-    "enzyme": "^2.7.1",
+    "enzyme": "^3.8.0",
+    "enzyme-adapter-react-16": "^1.9.1",
     "enzyme-to-json": "^1.5.0",
     "eslint": "^3.17.0",
     "eslint-config-airbnb": "^14.1.0",
@@ -38,10 +43,13 @@
     "eslint-plugin-react": "^6.10.0",
     "jest": "^19.0.2",
     "pre-commit": "^1.2.2",
-    "prop-types": "^15.6.0",
-    "react": "^15.4.2",
-    "react-addons-test-utils": "^15.4.2",
-    "react-dom": "^15.4.2"
+    "react": "^16.0.0",
+    "react-addons-test-utils": "^15.6.2",
+    "react-dom": "^16.0.0"
+  },
+  "peerDependencies": {
+    "react": "^15.3.0 || ^16.0.0",
+    "react-dom": "^15.3.0 || ^16.0.0"
   },
   "repository": {
     "type": "git",
